+++ app/javascript/dashboard/helper/CacheHelper/DataManager.js
